refactor(core): reorganiza pacotes e adiciona novos DTOs
Entidades movidas para model/entities para melhorar a estrutura do projeto. Criados novos DTOs: ProgramaDTO e AreaPesquisaDTO.
