+++ backend/src/main/java/com/acadmap/exception/EventoDuplicadoException.java
@@ -1,18 +1,18 @@
 package com.acadmap.exception;
 
 
-import com.acadmap.model.Evento;
+import com.acadmap.model.entities.Evento;
 import java.util.List;
 
 public class EventoDuplicadoException extends RuntimeException {
-    private final List<Evento> eventosSimilares;
+  private final List<Evento> eventosSimilares;
 
-    public EventoDuplicadoException(String message, List<Evento> eventosSimilares) {
-        super(message);
-        this.eventosSimilares = eventosSimilares;
-    }
+  public EventoDuplicadoException(String message, List<Evento> eventosSimilares) {
+    super(message);
+    this.eventosSimilares = eventosSimilares;
+  }
 
-    public List<Evento> getEventosSimilares() {
-        return eventosSimilares;
-    }
+  public List<Evento> getEventosSimilares() {
+    return this.eventosSimilares;
+  }
 }
