+++ backend/src/main/java/com/acadmap/model/dto/AreaPesquisaDTO.java
@@ -1,0 +1,7 @@
+package com.acadmap.model.dto;
+
+import java.util.UUID;
+
+public record AreaPesquisaDTO(UUID id, String nome) {
+
+}
