+++ backend/src/main/java/com/acadmap/model/dto/ProgramaDTO.java
@@ -1,0 +1,7 @@
+package com.acadmap.model.dto;
+
+import java.util.UUID;
+
+public record ProgramaDTO(UUID id, String nome) {
+
+}
