+++ backend/src/main/java/com/acadmap/repository/AreaPesquisaRepository.java
@@ -1,11 +1,7 @@
 package com.acadmap.repository;
 
-import com.acadmap.model.AreaPesquisa;
-
-import java.util.List;
+import com.acadmap.model.entities.AreaPesquisa;
 import java.util.UUID;
-
-import com.acadmap.model.Evento;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
