+++ backend/src/main/java/com/acadmap/repository/EventoRepository.java
@@ -1,9 +1,8 @@
 package com.acadmap.repository;
 
-import com.acadmap.model.Evento;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
+import com.acadmap.model.entities.Evento;
 import java.util.List;
 import java.util.UUID;
 
@@ -11,5 +10,5 @@ import java.util.UUID;
 public interface EventoRepository extends JpaRepository<Evento, UUID> {
 
 
-    List<Evento> findByNomeContainingIgnoreCase(String nome);
+  List<Evento> findByNomeContainingIgnoreCase(String nome);
 }
