+++ backend/src/main/java/com/acadmap/repository/ProgramaRepository.java
@@ -1,6 +1,6 @@
 package com.acadmap.repository;
 
-import com.acadmap.model.Programa;
+import com.acadmap.model.entities.Programa;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
