+++ backend/src/main/java/com/acadmap/repository/UsuarioRepository.java
@@ -1,6 +1,6 @@
 package com.acadmap.repository;
 
-import com.acadmap.model.Usuario;
+import com.acadmap.model.entities.Usuario;
 import java.util.Optional;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
