+++ backend/src/main/java/com/acadmap/service/UsuarioService.java
@@ -1,10 +1,10 @@
 package com.acadmap.service;
 
-import com.acadmap.model.AreaPesquisa;
-import com.acadmap.model.Programa;
-import com.acadmap.model.Usuario;
-import com.acadmap.model.dto.usuario.UsuarioRequestDTO;
-import com.acadmap.model.dto.usuario.UsuarioResponseDTO;
+import com.acadmap.model.dto.UsuarioRequestDTO;
+import com.acadmap.model.dto.UsuarioResponseDTO;
+import com.acadmap.model.entities.AreaPesquisa;
+import com.acadmap.model.entities.Programa;
+import com.acadmap.model.entities.Usuario;
 import com.acadmap.repository.AreaPesquisaRepository;
 import com.acadmap.repository.ProgramaRepository;
 import com.acadmap.repository.UsuarioRepository;
